Moved GeoSquare to model
