+++ app/src/main/java/pl/llp/aircasting/data/model/GeoSquare.kt
@@ -1,0 +1,7 @@
+package pl.llp.aircasting.data.model
+data class GeoSquare(
+    val north: Double,
+    val south: Double,
+    val east: Double,
+    val west: Double
+)
